Move ScanServerRefStoreImpl to server-base
